V3 update

+++ include/tfm_v3.hpp
@@ -343,7 +343,9 @@ class TFM
           }
         }
       }
+      
       //show();
+      
       return result;
     }
   public:
@@ -394,7 +396,7 @@ class TFM
     inline FIX               get_x_height() { return params[4];                                }
     inline FIX                get_em_size() { return params[5];                                }
     inline uint16_t       get_glyph_count() { return sizes.ec - sizes.bc + 1;                  }
-    inline uint8_t get_lig_kern_pgm_count() { return sizes.nl;                                 }
+    inline uint16_t get_lig_kern_pgm_count() { return sizes.nl;                                 }
     inline uint8_t         get_kern_count() { return sizes.nk;                                 }
     inline uint8_t   get_first_glyph_code() { return sizes.bc;                                 }
     inline uint8_t    get_last_glyph_code() { return sizes.ec;                                 }
